add creat file

+++ 0x15-file_io/1-create_file.c
@@ -29,7 +29,7 @@ int _strlen(char *s)
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	ssize_t bytes = 0, len = _strlen(text_content);
+	int bytes = 0, len = _strlen(text_content);
 
 	if (!filename)
 		return (-1);
